modulecmd: don't remove a loaded Pmodules module with module purge
Removing a Pmodule module with module purge has some unwanted
side-effects. It is not required to have modulecmd in a directory of
PATH. If you user explicitly loads a Pmodules module the bin directory
of the Pmodule module is added to the PATH. A Pmodules module might be
loaded for two reasons. First to change the Pmodules version, second
to make the build system available. If we remove a Pmodules module
with module purge the build systems is getting unavailable again. This
is confusing for the user.